Minor size improvements
git-svn-id: svn+ssh://svn.code.sf.net/p/jnode/svn/trunk@1899 407389f7-7c16-0410-9eee-ad85acb55d7f

+++ core/src/core/org/jnode/vm/x86/compiler/l1a/RefItem.java
@@ -50,7 +50,7 @@ final class RefItem extends WordItem implements X86CompilerConstants {
 	    super(factory);
 	}
 
-	final void initialize(EmitterContext ec, int kind, int offsetToFP, X86Register reg, VmConstString val) {
+	final void initialize(EmitterContext ec, byte kind, short offsetToFP, X86Register reg, VmConstString val) {
 		super.initialize(ec, kind, reg, offsetToFP);
 		this.value = val;
 	}
